requirements.txt
# ==============================================================================
# SIXBTC DEPENDENCIES
# ==============================================================================
# Python 3.11+ required
# Last updated: 2025-12-20

# Core Trading & Backtesting
# ==============================================================================
vectorbt==0.26.0                # Fast backtesting engine (1000x faster than Freqtrade)
pandas==2.2.0                   # Data manipulation
numpy==1.26.3                   # Numerical operations
ta-lib==0.4.28                  # Technical indicators (requires TA-Lib C library)
numba==0.59.0                   # JIT compilation for performance

# Hyperliquid Trading
# ==============================================================================
hyperliquid-python-sdk==0.2.4   # Official Hyperliquid SDK
eth-account==0.11.0             # Ethereum wallet (for HL authentication)
websockets==12.0                # WebSocket client for real-time data
aiohttp==3.9.1                  # Async HTTP client
requests==2.31.0                # Sync HTTP client (fallback)

# Binance (Historical Data)
# ==============================================================================
ccxt==4.2.25                    # Binance historical data download
python-binance==1.0.19          # Binance API wrapper (alternative)

# Database (PostgreSQL)
# ==============================================================================
psycopg2-binary==2.9.9          # PostgreSQL driver
sqlalchemy==2.0.25              # ORM and query builder
alembic==1.13.1                 # Database migrations

# AI Providers (Strategy Generation)
# ==============================================================================
anthropic==0.18.0               # Claude API (Anthropic)
google-generativeai==0.3.2      # Gemini API (Google)
openai==1.10.0                  # GPT/Codex API (OpenAI)

# Configuration & Environment
# ==============================================================================
pyyaml==6.0.1                   # YAML config parser
python-dotenv==1.0.0            # .env file loader
pydantic==2.5.3                 # Config validation with type hints
pydantic-settings==2.1.0        # Settings management

# Scheduling & Orchestration
# ==============================================================================
apscheduler==3.10.4             # Multi-timeframe scheduling
asyncio-mqtt==0.16.2            # Async event loop utilities (if needed)

# Monitoring & Logging
# ==============================================================================
rich==13.7.0                    # Beautiful TUI dashboard
loguru==0.7.2                   # Enhanced logging (alternative to stdlib)
prometheus-client==0.19.0       # Metrics export (optional)

# CLI & User Interface
# ==============================================================================
click==8.1.7                    # CLI framework
typer==0.9.0                    # Modern CLI with type hints (alternative to click)
tabulate==0.9.0                 # Pretty-print tables

# Utilities
# ==============================================================================
python-dateutil==2.8.2          # Date parsing
pytz==2024.1                    # Timezone handling
tqdm==4.66.1                    # Progress bars

# Testing
# ==============================================================================
pytest==7.4.4                   # Testing framework
pytest-asyncio==0.23.3          # Async test support
pytest-cov==4.1.0               # Coverage reporting
pytest-mock==3.12.0             # Mocking utilities

# Code Quality (Development)
# ==============================================================================
black==24.1.1                   # Code formatter
flake8==7.0.0                   # Linter
mypy==1.8.0                     # Static type checker
isort==5.13.2                   # Import sorter

# Performance Profiling (Optional)
# ==============================================================================
line-profiler==4.1.1            # Line-by-line profiling
memory-profiler==0.61.0         # Memory usage profiling

# ==============================================================================
# INSTALLATION NOTES
# ==============================================================================
#
# 1. TA-Lib requires C library installation first:
#    Ubuntu/Debian: sudo apt-get install ta-lib
#    macOS: brew install ta-lib
#    Or build from source: http://ta-lib.org/hdr_dw.html
#
# 2. PostgreSQL required (via Docker):
#    docker-compose up -d postgres
#
# 3. Install all dependencies:
#    pip install -r requirements.txt
#
# 4. For development (includes testing + code quality):
#    pip install -r requirements.txt
#
